# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: Threavia Core
  # Filled in by the server with the version of the running build, which the
  # linker wrote into the binary. A release number kept in this file would be a
  # line someone has to remember to change, and the one that drifts from what is
  # actually serving.
  version: "0.0.0"
  description: |
    HTTP/JSON for commands, queries, snapshots and history; SSE for realtime
    events. No GraphQL and no WebSocket in V1.

    Every user-scoped route is checked against the authenticated caller:
    `404` covers both "does not exist" and "not yours", so an identifier cannot
    be probed for existence.

    Requests a page of another site could send are refused before they are
    handled: a JSON body not declared as `application/json` gets `415
    unsupported_media_type`, a write a browser marks as cross-site gets `403
    forbidden_origin`, and a `Host` outside `THREAVIA_HTTP_ALLOWED_HOSTS` gets
    `421 misdirected_request`. A client that is not a browser sends neither
    `Origin` nor `Sec-Fetch-Site` and is unaffected.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /
    description: The Core instance serving this document
...
security:
  - basicAuth: []
  - bearerAuth: []
paths:
...
  /api/v1/backends/register:
    post:
      tags: [backends]
      summary: Register a BackendInstance
      description: |
        The one route a backend calls before it has any user credential. It is
        gated by a one-shot registration token or by the shared registration
        key, never by user authentication, and returns the persistent
        credential once.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, capabilities]
              properties:
                name: { type: string }
                capabilities:
                  type: array
                  items: { $ref: "#/components/schemas/Capability" }
                token:
                  type: string
                  description: A one-shot user token, which owns the instance immediately.
                sharedKey:
                  type: string
                  description: The shared key, which creates an UNCLAIMED instance plus a claim code.
      responses:
        "201":
          description: The instance and its persistent credential
          content:
            application/json:
              schema:
                type: object
                required: [backendInstanceId, credential, claimed]
                properties:
                  backendInstanceId: { $ref: "#/components/schemas/UUID" }
                  credential: { type: string }
                  claimed: { type: boolean }
                  claimCode:
                    type: string
                    description: Present only when the instance is UNCLAIMED.
        "403":
          $ref: "#/components/responses/Error"
...
  /api/v1/backends/{backendId}:
    parameters:
      - $ref: "#/components/parameters/BackendID"
    get:
      tags: [backends]
      summary: One BackendInstance
      responses:
        "200":
          description: The instance
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BackendInstance" }
        "404":
          $ref: "#/components/responses/Error"
  /api/v1/backends/{backendId}/revoke:
    parameters:
      - $ref: "#/components/parameters/BackendID"
    post:
      tags: [backends]
      summary: Revoke the persistent credential
      description: The record stays for historical references, and a returning backend must register anew.
      responses:
        "204":
          description: Revoked
        "404":
          $ref: "#/components/responses/Error"
...
  /api/v1/backends/claim:
    post:
      tags: [backends]
      summary: Claim an UNCLAIMED instance with its one-time code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [claimCode]
              properties:
                claimCode: { type: string }
      responses:
        "200":
          description: The claimed instance
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BackendInstance" }
        "403":
          $ref: "#/components/responses/Error"
  /api/v1/backend-tokens:
    post:
      tags: [backends]
      summary: Issue a one-shot registration token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label: { type: string }
                ttlSeconds: { type: integer, format: int64 }
      responses:
        "201":
          description: The token, returned once
          content:
            application/json:
              schema:
                type: object
                required: [token, expiresAt]
                properties:
                  token: { type: string }
                  expiresAt: { type: string, format: date-time }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Used when THREAVIA_AUTH_MODE is basic.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        An OIDC token, used when THREAVIA_AUTH_MODE is oidc. Core runs no login
        flow: a client obtains its token from the provider.
  parameters:
...
    BackendID:
      name: backendId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/UUID" }
...
  responses:
    Error:
      description: The single error shape of the client API
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code: { type: string }
                  message: { type: string }
  schemas:
    UUID:
      type: string
      format: uuid
...
    Capability:
      type: string
      enum: [CODE, INTERACTION, REVIEW]
...
    BackendOwnershipStatus:
      type: string
      enum: [UNCLAIMED, CLAIMED, REVOKED]
    BackendOperationalStatus:
      type: string
      description: BUSY is deliberately absent; capacity is modelled separately.
      enum: [READY, DEGRADED, OFFLINE, UNAVAILABLE]
    ProviderAuthState:
      type: string
      enum: [UNKNOWN, AUTHENTICATED, AUTHENTICATION_REQUIRED]
...
    Capacity:
      type: object
      required: [maxConcurrentRuns, activeRuns]
      properties:
        maxConcurrentRuns: { type: integer }
        activeRuns: { type: integer }
    Condition:
      type: object
      description: Why a backend is in the state it reports, in the vocabulary the backend chose. Core stores it without interpreting it.
      required: [type, status, observedAt]
      properties:
        type: { type: string, examples: ["ProviderAvailable"] }
        status: { type: string, examples: ["False"] }
        reason: { type: string, examples: ["ExecutableNotFound"] }
        message: { type: string }
        observedAt: { type: string, format: date-time }
    BackendInstance:
      type: object
      description: An autonomous remote execution participant. It belongs to a User, not to a Project.
      required: [id, name, ownershipStatus, operationalStatus, providerAuthState, capabilities, capacity, protocolVersion, createdAt, updatedAt]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        ownerId:
          type: string
          description: Absent while the instance is UNCLAIMED.
        name: { type: string }
        ownershipStatus: { $ref: "#/components/schemas/BackendOwnershipStatus" }
        operationalStatus: { $ref: "#/components/schemas/BackendOperationalStatus" }
        providerAuthState: { $ref: "#/components/schemas/ProviderAuthState" }
        capabilities:
          type: array
          items: { $ref: "#/components/schemas/Capability" }
        features:
          type: array
          description: |
            The optional CODE features announced on the last connection (spec
            sections 3.5 and 7): whether a message can reach a Job while it
            runs, after interrupting it or at its next step.
          items:
            type: string
            enum: [JOB_INPUT_NOW, JOB_INPUT_NEXT]
        capacity: { $ref: "#/components/schemas/Capacity" }
        conditions:
          type: array
          items: { $ref: "#/components/schemas/Condition" }
        protocolVersion: { type: integer }
        lastHeartbeatAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        revokedAt: { type: string, format: date-time }
...